requirements.txt
psycopg2-binary
faker
python-dotenv
psycopg2